pyproject.toml
[tool.poetry]
name = "shell"
version = "0.1.0"
description = ""
authors = ["Ashish Parimoo <ashish.parimoo@anptech.co.uk>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
black = "^23.3.0"
psycopg2-binary = "^2.9.6"
SQLAlchemy = "^2.0.15"
python-dateutil = "^2.8.2"
scipy = "^1.10.1"
scikit-learn = "^1.2.2"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

